+++ src/main/kotlin/one/oktw/galaxy/block/event/BlockEvents.kt
@@ -29,7 +29,7 @@ import net.minecraft.util.math.Direction
 import net.minecraft.world.GameMode
 import one.oktw.galaxy.block.Block
 import one.oktw.galaxy.block.type.BlockType
-import one.oktw.galaxy.block.util.CustomBlockUtil
+import one.oktw.galaxy.block.util.BlockUtil
 import one.oktw.galaxy.event.annotation.EventListener
 import one.oktw.galaxy.event.type.BlockBreakEvent
 import one.oktw.galaxy.event.type.PlayerInteractBlockEvent
@@ -37,7 +37,6 @@ import one.oktw.galaxy.event.type.PlayerInteractItemEvent
 import one.oktw.galaxy.item.Tool
 import one.oktw.galaxy.item.type.ItemType
 import one.oktw.galaxy.item.type.ToolType
-import one.oktw.galaxy.player.util.BlockUtil
 
 class BlockEvents {
     private val eventLock = HashSet<PlayerInteractBlockC2SPacket>()
@@ -55,31 +54,16 @@ class BlockEvents {
     @Suppress("unused")
     @EventListener(true)
     fun onPlayerInteractBlock(event: PlayerInteractBlockEvent) {
-        val player = event.player
-        val hand = event.packet.hand
-        val hitResult = event.packet.blockHitResult
-        val blockPos = hitResult.blockPos
-
-        if (!player.isSneaking && BlockUtil.isMature(player.serverWorld, blockPos, player.serverWorld.getBlockState(blockPos))) return
-
-        val tryUseBlock = CustomBlockUtil.vanillaTryUseBlock(player, hand, hitResult)
-        if (tryUseBlock.isAccepted) {
-            event.cancel = true
-            if (tryUseBlock.shouldSwingHand()) player.swingHand(hand, true)
-            return
-        }
-
-
-        if (eventLock.contains(event.packet) || usedLock.contains(player)) return
+        if (eventLock.contains(event.packet) || usedLock.contains(event.player)) return
         eventLock.add(event.packet)
 
         var finished: Boolean
-        finished = tryBreakBlock(event.packet, player, hand)
+        finished = tryBreakBlock(event.packet, event.player, event.packet.hand)
         if (!finished) finished = tryOpenGUI(event)
-        if (!finished) finished = tryPlaceBlock(event.packet, player)
+        if (!finished) finished = tryPlaceBlock(event.packet, event.player)
         if (finished) {
-            player.swingHand(hand, true)
-            usedLock.add(player)
+            event.player.swingHand(event.packet.hand, true)
+            usedLock.add(event.player)
             event.cancel = true
         }
     }
@@ -94,7 +78,7 @@ class BlockEvents {
     @EventListener(true)
     fun onBlockBreak(event: BlockBreakEvent) {
         if (event.state.block != Blocks.BARRIER) return
-        CustomBlockUtil.getCustomBlockEntity((event.world as ServerWorld), event.pos) ?: return
+        BlockUtil.getCustomBlockEntity((event.world as ServerWorld), event.pos) ?: return
         event.cancel = true
     }
 
@@ -103,28 +87,28 @@ class BlockEvents {
         val position = packet.blockHitResult.blockPos
         if (world.getBlockState(position).block != Blocks.BARRIER) return false
         if (player.isSneaking && player.getStackInHand(hand).isItemEqual(Tool(ToolType.WRENCH).createItemStack())) {
-            CustomBlockUtil.getCustomBlockEntity(world, position) ?: return false
-            CustomBlockUtil.removeBlock(world, position)
+            BlockUtil.getCustomBlockEntity(world, position) ?: return false
+            BlockUtil.removeBlock(world, position)
             return true
         }
         return false
     }
 
     private fun tryOpenGUI(event: PlayerInteractBlockEvent): Boolean {
-        val player = event.player
-        val world = player.serverWorld
+        val world = event.player.serverWorld
         val position = event.packet.blockHitResult.blockPos
         val hand = event.packet.hand
-        if (!player.shouldCancelInteraction()) {
-            val entity = CustomBlockUtil.getCustomBlockEntity(world, position) ?: return false
-            val blockType = CustomBlockUtil.getTypeFromCustomBlockEntity(entity) ?: return false
-            if (blockType.hasGUI && hand == Hand.MAIN_HAND) openGUI(blockType, player)
+        if (!event.player.isSneaking) {
+            val entity = BlockUtil.getCustomBlockEntity(world, position) ?: return false
+            val blockType = BlockUtil.getTypeFromCustomBlockEntity(entity) ?: return false
+            if (blockType.hasGUI && hand == Hand.MAIN_HAND) openGUI(blockType, event.player, event)
             return blockType.hasGUI
         }
         return false
     }
 
-    private fun openGUI(blockType: BlockType, player: ServerPlayerEntity) {
+    private fun openGUI(blockType: BlockType, player: ServerPlayerEntity, event: PlayerInteractBlockEvent) {
+        event.cancel = true
         when (blockType) { // TODO activate GUI
             BlockType.CONTROL_PANEL -> player.sendMessage(LiteralText("Control Panel"), false)
             BlockType.PLANET_TERMINAL -> player.sendMessage(LiteralText("Planet Terminal"), false)
@@ -139,7 +123,7 @@ class BlockEvents {
         val world = player.serverWorld
         val server = player.server
         val position = packet.blockHitResult.blockPos
-        val placePosition = CustomBlockUtil.getPlacePosition(world, position, packet.blockHitResult)
+        val placePosition = BlockUtil.getPlacePosition(world, position, packet.blockHitResult)
 
         val hand = packet.hand
