+++ src/main/kotlin/one/oktw/galaxy/block/util/BlockUtil.kt
@@ -1,6 +1,6 @@
 /*
  * OKTW Galaxy Project
- * Copyright (C) 2018-2019
+ * Copyright (C) 2018-2020
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published
@@ -27,27 +27,20 @@ import net.minecraft.entity.TntEntity
 import net.minecraft.entity.vehicle.AbstractMinecartEntity
 import net.minecraft.entity.vehicle.BoatEntity
 import net.minecraft.item.ItemStack
-import net.minecraft.item.ItemUsageContext
-import net.minecraft.server.network.ServerPlayerEntity
 import net.minecraft.server.world.ServerWorld
 import net.minecraft.sound.SoundCategory
 import net.minecraft.sound.SoundEvents
 import net.minecraft.state.property.Properties
-import net.minecraft.util.ActionResult
-import net.minecraft.util.Hand
 import net.minecraft.util.hit.BlockHitResult
 import net.minecraft.util.math.BlockPos
 import net.minecraft.util.math.Box
-import net.minecraft.world.GameMode
 import one.oktw.galaxy.block.Block
 import one.oktw.galaxy.block.type.BlockType
-import one.oktw.galaxy.item.type.ItemType
 import net.minecraft.block.Block as minecraftBlock
 
-object CustomBlockUtil {
+object BlockUtil {
     fun placeAndRegisterBlock(world: ServerWorld, blockPos: BlockPos, blockItem: ItemStack, blockType: BlockType): Boolean {
         val entities = world.getEntities(null, Box(blockPos))
-        // block place entity blacklist
         if (entities.any { entity -> entity is LivingEntity || entity is BoatEntity || entity is AbstractMinecartEntity || entity is TntEntity }) {
             return false
         }
@@ -104,41 +97,6 @@ object CustomBlockUtil {
 
     fun getPlacePosition(world: ServerWorld, blockPos: BlockPos, blockHitResult: BlockHitResult): BlockPos =
         if (isReplaceable(world, blockPos)) blockPos else blockPos.offset(blockHitResult.side)
-
-    fun vanillaTryUseBlock(player: ServerPlayerEntity, hand: Hand, hitResult: BlockHitResult): ActionResult {
-        // Vanilla
-        if (player.interactionManager.gameMode != GameMode.SPECTATOR) {
-            if (!((!player.mainHandStack.isEmpty || !player.offHandStack.isEmpty) && player.shouldCancelInteraction())) {
-                val world = player.serverWorld
-                val tryUseBlock = world.getBlockState(hitResult.blockPos).onUse(world, player, hand, hitResult)
-                if (tryUseBlock.isAccepted) {
-                    return tryUseBlock
-                } else {
-                    // Prevent duplicate
-                    val itemStack = player.getStackInHand(hand)
-                    return if (!itemStack.isEmpty && !player.itemCooldownManager.isCoolingDown(itemStack.item)) {
-                        val itemUsageContext = ItemUsageContext(player, hand, hitResult)
-                        // Vanilla stops here and check if it is custom block item first
-                        val tag = itemStack.tag
-                        val itemType = tag?.getString("customItemType")
-                        if (itemType == ItemType.BLOCK.name) return ActionResult.PASS
-                        // Vanilla continue
-                        if (player.interactionManager.isCreative) {
-                            val count = itemStack.count
-                            val tryUseOnBlock = itemStack.useOnBlock(itemUsageContext)
-                            itemStack.count = count
-                            tryUseOnBlock
-                        } else {
-                            itemStack.useOnBlock(itemUsageContext)
-                        }
-                    } else {
-                        ActionResult.PASS
-                    }
-                }
-            }
-        }
-        return ActionResult.PASS
-    }
 
     private fun isReplaceable(world: ServerWorld, blockPos: BlockPos) = when (world.getBlockState(blockPos).block) {
         is TallFlowerBlock -> false
